### 2023-04-22 #5396 @editorialbot
[PRE REVIEW]: G2Aero: A Python package for separable shape tensors
**Submitting author:** @olgadoronina (<a href="http://orcid.org/0000-0003-3413-3197">Olga Doronina</a>)
**Repository:** https://github.com/NREL/G2Aero
**Branch with paper.md** (empty if default branch): main
**Version:** v0.1.2
