### 2023-04-22 #5396 @danielskatz
👋 @olgadoronina

I think https://github.com/NREL/G2Aero/pull/1 will fix the bibtex parsing error. Please merge this, then use the command `@editorialbot check references` to see if it does, and to check to the references more generally. editorialbot commands need to be the first entry in a new comment.
