### 2023-04-23 #5396 @isdanni
> @isdanni - sorry, your github name isn't on our editor list, as far as I can tell, so I'm confused...

I actually just joined the community and added myself through the [Volunteer to Review](https://reviewers.joss.theoj.org/join) link, so this is just me checking to see if any reviewer is needed. Totally understand if people with more senior review experience are preferred :)
