+++ paper/paper.bib
@@ -160,13 +160,12 @@
 
 
 @misc{wisdem,
-	title = {WISDEM® (Wind-Plant Integrated System Design and Engineering Model), Version v3.4.1},
+	title = {WISDEM{\textregistered} (Wind-Plant Integrated System Design and Engineering Model), Version v3.4.1},
 	author = {Dykes, Katherine and Ning, S. Andrew and Scott, George and Graf, Peter and USDOE Office of Energy Efficiency and Renewable Energy},
 	doi = {10.11578/dc.20211208.2},
 	url = {https://www.osti.gov/biblio/1834705}, 
 	year = {2021},
-	month = {2},
-	note =
+	month = {2}
 }
 
 @misc{BigFoil:2022,
