+++ paper/paper.md
@@ -78,7 +78,7 @@ Specific methods within `G2Aero` include:
 `G2Aero` defines an improved parameter domain inferred from a database of relevant shapes (discrete 2D curves). This domain independently treats affine and undulating type deformations, allowing for more targeted shape design and generation of a rich set of novel 2D deformations. Using PGA over the Grassmannian significantly reduces the dimensionality of the parameter domain. Using an extensive database of airfoils [@BigFoil:2022] and analyzing the shape reconstruction error, we found that we can use as few as four PGA parameters to represent a wide range of undulation-type deformations of the shape and two parameters to represent linear deformations.
 
 In @Doronina:2022, we concisely summarized the framework and demonstrated the advantages of an improved parameter domain for ML/AI algorithms.
-@zhang2022aerodynamic tested Grassmannian shape representation and demonstrated robustness for shape optimization applications. @jasa2022wind used airfoils generated by `G2Aero` coupled with NREL’s Wind-Plant Integrated System Design Engineering Model (WISDEM&#174;) [@wisdem] to design blade shapes with reduced costs of energy compared to traditional design methods.
+@zhang2022aerodynamic tested Grassmannian shape representation and demonstrated robustness for shape optimization applications. @jasa2022wind used airfoils generated by `G2Aero` coupled with NREL’s Wind-Plant Integrated System Design Engineering Model (WISDEM) [@wisdem] to design blade shapes with reduced costs of energy compared to traditional design methods.
 
 ### 2) Building 3D blades by interpolating 2D airfoil cross-sections
 
@@ -95,6 +95,8 @@ We use piecewise geodesic interpolation to connect a sequence of 2D cross-sectio
 Despite our focus on airfoil and blade design, we note obvious extensions to a variety of 2D shape applications as well as 3D shapes that are well-defined by a sequence of cross-sections, such as inlets, nozzles, ducts, passages, channels, etc. We encourage users to experiment with `G2Aero` and apply methods to alternative types of shapes and surfaces.  Our computationally efficient approach to matrix-manifold learning and generative modeling of discrete planar curves may offer advantages in applications beyond aerodynamics, where learning a non-linear topology of shapes from data may be impactful.
 
 # Acknowledgments
+
+Certain software are identified in this paper in order to specify the experimental procedure adequately.  Such identification is not intended to imply recommendation or endorsement of any product or service by NIST, nor is it intended to imply that the software identified are necessarily the best available for the purpose.
 
 This work was authored in part by the National Renewable Energy Laboratory, operated by Alliance for Sustainable Energy, LLC, for the U.S. Department of Energy (DOE) under Contract No. DE-AC36-08GO28308. Funding is partially provided by the Advanced Research Projects Agency-Energy (ARPA-E) Design Intelligence Fostering Formidable Energy Reduction and Enabling Novel Totally Impactful Advanced Technology Enhancements (DIFFERENTIATE) program. The views expressed in the article do not necessarily represent the views of the DOE or the U.S. Government. This work is U.S. Government work and not protected by U.S. copyright. A portion of this research was performed using computational resources sponsored by the Department of Energy's Office of Energy Efficiency and Renewable Energy and located at the National Renewable Energy Laboratory.
 
