### 2023-04-24 #5396 @olgadoronina
I had to update the trademark disclaimer in the paper Acknowledgement section to satisfy NIST requirements. See [latest commit](https://github.com/NREL/G2Aero/commit/3b1ef384599dcbc47998b22d1a1ae9a8684af9d5)
