### 2023-05-12 f14b4df
Adding missing Function Explanation in API Doc
