+++ src/g2aero/reparametrization.py
@@ -57,8 +57,8 @@ def get_landmarks(xy, n_landmarks=401, method='planar', add_gap=False, **kwargs)
 
 
 def planar_reparametrization(xy, n_landmarks, sampling='uniform', **kwargs):
-    """ Builds cubic splines of x(t) and y(t) where t in [0, 1] is coordinate 
-    along the arclength. Resamples landmark 
+    """Builds cubic splines of x(t) and y(t) where t in [0, 1] is coordinate 
+    along the arclength. Resamples landmark.
 
     :param xy: (n, 2) given coordinated defining the shape
     :param n_landmarks: scalar number of landmarks in returned shape
@@ -135,6 +135,16 @@ def planar_reparametrization(xy, n_landmarks, sampling='uniform', **kwargs):
 
 
 def polar_reparametrization(xy, n_landmarks=401, sampling='uniform_gr', **kwargs):
+    """Builds cubic splines of polar coordinates theta(t) and alpha(t), 
+    where t in [0, 1] is coordinate along the arclength. Resamples landmark.
+
+    :param xy: (n, 2) given coordinated defining the shape
+    :param n_landmarks: scalar number of landmarks in returned shape, defaults to 401
+    :param sampling: distribution of landmarks along the arclength 
+                     possible options: 'uniform_phys', 'uniform_gr' and 'curvature', 
+                     defaults to 'uniform_gr'
+    :return: (n_landmarks, 2) array of landmarks after reparametrization
+    """
 
     xy_gr, M, b = landmark_affine_transform(xy)
     t = arc_distance(xy_gr)     # arc length
@@ -167,6 +177,15 @@ def polar_reparametrization(xy, n_landmarks=401, sampling='uniform_gr', **kwargs
 
 
 def cst_reparametrization(xy, n_landmarks=401, original_landmarks=False, name='', cst_order=8, **kwargs):
+    """Provides landmarks after CST reparametrization.
+
+    :param xy: (n, 2) given coordinated defining the shape
+    :param n_landmarks: scalar number of landmarks in returned shape, defaults to 401
+    :param original_landmarks: (bool) use landmarks from given shape coordinate xy, defaults to False
+    :param name: (str) name of airfoil used to define normals in CST representation (using n1, n2 = 0.5, 0.5 for circles), defaults to ''
+    :param cst_order:  order of the polynomial used, defaults to 8
+    :return: (n_landmarks, 2) array of landmarks after reparametrization
+    """
     xy = np.asarray(xy)
     if name in ['circular', 'Cylinder', 'Cylinder1', 'Cylinder2']:
         n1, n2 = 0.5, 0.5
@@ -207,8 +226,7 @@ def cst_reparametrization(xy, n_landmarks=401, original_landmarks=False, name=''
 
 
 def calc_cst_param(x, y, n1, n2, y_tailedge=0.0, order=8):
-    """
-    Solve the least squares problem for a given shape
+    """Solve the least squares problem for a given shape.
 
     :param x: (np.ndarray): (x/c) coordinates locations
     :param y: (np.ndarray): (y/c) coordinate locations
@@ -224,6 +242,15 @@ def calc_cst_param(x, y, n1, n2, y_tailedge=0.0, order=8):
 
 
 def cst_matrix(x, n1, n2, order):
+    """Construct CST matrix A used in CST representation
+    y_{low/upp} = A coef_CST + b 
+   
+    :param x: x/c coordinates locations
+    :param n1: normal coord
+    :param n2: normal coord
+    :param order: order of the polynomial used
[…]
