+++ docs/community.rst
@@ -1,5 +1,4 @@
 Feedback, Support, and Contributions
 ====================================
 
-Contributions are always welcome! 
-To contribute to G2aero, report an issue, or seek support, please initiate a pull request or issue through the project `github <https://github.com/NREL/G2Aero>`_.
+.. mdinclude:: ../../CONTRIBUTING.md
