+++ tests/test_Grassmann.py
@@ -1,840 +1,43 @@
+import os
 from unittest import TestCase
 from g2aero.Grassmann import *
-
-X_phys = [[ 1.00000800e+00, -7.10000000e-04],
-       [ 9.94971356e-01, -5.42445001e-04],
-       [ 9.89935193e-01, -3.61225537e-04],
-       [ 9.84899323e-01, -1.71780602e-04],
-       [ 9.79863559e-01,  2.05133067e-05],
-       [ 9.74827778e-01,  2.12227161e-04],
-       [ 9.69791934e-01,  4.02213649e-04],
-       [ 9.64755989e-01,  5.89519231e-04],
-       [ 9.59719930e-01,  7.73763435e-04],
-       [ 9.54683756e-01,  9.54868175e-04],
-       [ 9.49647470e-01,  1.13282479e-03],
-       [ 9.44611084e-01,  1.30789654e-03],
-       [ 9.39574609e-01,  1.48041497e-03],
-       [ 9.34538055e-01,  1.65058198e-03],
-       [ 9.29501419e-01,  1.81834157e-03],
-       [ 9.24464701e-01,  1.98360259e-03],
-       [ 9.19427895e-01,  2.14616164e-03],
-       [ 9.14390992e-01,  2.30569998e-03],
-       [ 9.09353987e-01,  2.46195304e-03],
-       [ 9.04316889e-01,  2.61519501e-03],
-       [ 8.99279717e-01,  2.76598696e-03],
-       [ 8.94242483e-01,  2.91471773e-03],
-       [ 8.89205180e-01,  3.06104769e-03],
-       [ 8.84167793e-01,  3.20444904e-03],
-       [ 8.79130313e-01,  3.34459521e-03],
-       [ 8.74092746e-01,  3.48156652e-03],
-       [ 8.69055097e-01,  3.61549041e-03],
-       [ 8.64017371e-01,  3.74644450e-03],
-       [ 8.58979568e-01,  3.87445455e-03],
-       [ 8.53941692e-01,  3.99952553e-03],
-       [ 8.48903739e-01,  4.12149390e-03],
-       [ 8.43865707e-01,  4.24010545e-03],
-       [ 8.38827592e-01,  4.35517153e-03],
-       [ 8.33789399e-01,  4.46678221e-03],
-       [ 8.28751135e-01,  4.57510010e-03],
-       [ 8.23712802e-01,  4.68019726e-03],
-       [ 8.18674401e-01,  4.78196215e-03],
-       [ 8.13635931e-01,  4.88025947e-03],
-       [ 8.08597391e-01,  4.97491654e-03],
-       [ 8.03558780e-01,  5.06572177e-03],
-       [ 7.98520099e-01,  5.15250495e-03],
-       [ 7.93481354e-01,  5.23548929e-03],
-       [ 7.88442554e-01,  5.31510834e-03],
-       [ 7.83403706e-01,  5.39166452e-03],
-       [ 7.78364810e-01,  5.46490912e-03],
-       [ 7.73325861e-01,  5.53445169e-03],
-       [ 7.68286858e-01,  5.60000664e-03],
-       [ 7.63247805e-01,  5.66149676e-03],
-       [ 7.58208702e-01,  5.71887120e-03],
-       [ 7.53169554e-01,  5.77212050e-03],
-       [ 7.48130365e-01,  5.82127712e-03],
-       [ 7.43091138e-01,  5.86638687e-03],
-       [ 7.38051878e-01,  5.90759588e-03],
-       [ 7.33012589e-01,  5.94510217e-03],
-       [ 7.27973274e-01,  5.97902682e-03],
[…]
