### 2023-07-08 #5408 @fdmatoz
Review done. Just a couple of suggestion to the authors, Fix the warning in python 3.11:

=============================== warnings summary ===============================
../../../miniconda3/envs/g2aero/lib/python3.11/site-packages/g2aero-0.1.1-py3.11.egg/g2aero/Grassmann.py:56
../../../miniconda3/envs/g2aero/lib/python3.11/site-packages/g2aero-0.1.1-py3.11.egg/g2aero/Grassmann.py:56
  /home/dd/miniconda3/envs/g2aero/lib/python3.11/site-packages/g2aero-0.1.1-py3.11.egg/g2aero/Grassmann.py:56: DeprecationWarning: invalid escape sequence '\D'

../../../miniconda3/envs/g2aero/lib/python3.11/site-packages/g2aero-0.1.1-py3.11.egg/g2aero/Grassmann.py:70
../../../miniconda3/envs/g2aero/lib/python3.11/site-packages/g2aero-0.1.1-py3.11.egg/g2aero/Grassmann.py:70
  /home/dd/miniconda3/envs/g2aero/lib/python3.11/site-packages/g2aero-0.1.1-py3.11.egg/g2aero/Grassmann.py:70: DeprecationWarning: invalid escape sequence '\D'

../../../miniconda3/envs/g2aero/lib/python3.11/site-packages/g2aero-0.1.1-py3.11.egg/g2aero/SPD.py:103
../../../miniconda3/envs/g2aero/lib/python3.11/site-packages/g2aero-0.1.1-py3.11.egg/g2aero/SPD.py:103
  /home/dd/miniconda3/envs/g2aero/lib/python3.11/site-packages/g2aero-0.1.1-py3.11.egg/g2aero/SPD.py:103: DeprecationWarning: invalid escape sequence '\i'

../../../miniconda3/envs/g2aero/lib/python3.11/site-packages/g2aero-0.1.1-py3.11.egg/g2aero/SPD.py:119
../../../miniconda3/envs/g2aero/lib/python3.11/site-packages/g2aero-0.1.1-py3.11.egg/g2aero/SPD.py:119
  /home/dd/miniconda3/envs/g2aero/lib/python3.11/site-packages/g2aero-0.1.1-py3.11.egg/g2aero/SPD.py:119: DeprecationWarning: invalid escape sequence '\D'

-- Docs: https://docs.pytest.org/en/stable/how-to/capture-warnings.html
================== 14 passed, 8 warnings in 89.91s (0:01:29) ===================

Docs: Provide with example on how to create the figures presented in the paper. Sometimes this is overlooked but is important for people that are getting into the usage of the package with other models than the ones presented in the paper.
