### 2023-07-17 #5408 @olgadoronina
Thank you @isdanni and @fdmatoz for your review and comments.

@fdmatoz thanks for your suggestions, we will fix the warnings in Python 3.11. Most of the figures in the paper are taken from the tutorials in the docs or provided example scripts, but we will add more directions in tutorials to make it easy to reproduce the paper figures.

@vissarion, it seems like all checkboxes are checked. Is there anything else we should do to finalize the process?
