### 2023-07-17 04b4ca9
fix parsing .md files in docs
