+++ docs/conf.py
@@ -31,10 +31,27 @@ release = 'v0.1.2'
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
 extensions = ['nbsphinx_link', 'nbsphinx', 
-              'sphinx.ext.autodoc', 'sphinx.ext.autosummary', 'sphinx.ext.coverage', 'sphinx.ext.napoleon', 'sphinx.ext.autosectionlabel',
-              'sphinxcontrib.bibtex']
+              'sphinx.ext.autodoc', 'sphinx.ext.autosummary', 'sphinx.ext.coverage', 
+              'sphinx.ext.napoleon', 'sphinx.ext.autosectionlabel',
+              'sphinxcontrib.bibtex', "myst_parser"]
 
 bibtex_bibfiles = ['refs.bib']
+
+myst_enable_extensions = [
+    "amsmath",
+    "attrs_inline",
+    "colon_fence",
+    "deflist",
+    "dollarmath",
+    "fieldlist",
+    "html_admonition",
+    "html_image",
+    "replacements",
+    "smartquotes",
+    "strikethrough",
+    "substitution",
+    "tasklist",
+]
 
 autodoc_mock_imports = ['numpy', 'scipy', 'gmsh', 'g2aero']
 # Add any paths that contain templates here, relative to this directory.
