### 2023-08-02 #5408 @olgadoronina
@vissarion, it seems like both reviewers confirmed that they finished their review and recommended acceptance. Is anything else we should do?
