### 2023-08-04 #5408 @vissarion
When a submission is ready to be accepted, we ask that the authors issue a new tagged release of the software (if changed), and archive it (see [this guide](https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content)). Please do this and post the version number and archive DOI here.
