### 2023-09-19 #5408 @vissarion
Thanks @olgadoronina the comment "JOSS submission" is not needed since this will eventually become a published paper :)
