### 2023-09-19 NREL/G2Aero#6 @vissarion
JOSS submission
Some minor comments on references.

In "Quayle, M. (2022). BigFoil.com airfoil database." the website could be a clickable link.

In " Chen, W. et al, 2020" bézier should start with a capital letter.
