### 2023-09-19 #5408 @vissarion
@olgadoronina almost ready to recommend acceptance. I have just some minor comments in https://github.com/NREL/G2Aero/issues/6
