### 2023-09-19 23bebad
minor update to paper.bib
