+++ paper/paper.bib
@@ -24,7 +24,7 @@
 
 
 @article{chen2020airfoil,
-	title={Airfoil design parameterization and optimization using b{\'e}zier generative adversarial networks},
+	title={Airfoil design parameterization and optimization using {B}{\'e}zier generative adversarial networks},
 	author={Chen, Wei and Chiu, Kevin and Fuge, Mark D},
 	journal={AIAA journal},
 	volume={58},
@@ -183,9 +183,9 @@
 
 @misc{BigFoil:2022,
 	author = {Quayle, Michael},
-	title = {BigFoil.com Airfoil Database},
+	title = {\url{BigFoil.com} Airfoil Database},
 	year = {2022},
-	note = { \url{https://www.bigfoil.com}},
+	note = {\url{https://www.bigfoil.com}},
 }
 %%%% use our representation
 
