+++ paper/paper.bib
@@ -183,8 +183,9 @@
 
 @misc{BigFoil:2022,
 	author = {Quayle, Michael},
-	title = {\url{BigFoil.com} Airfoil Database},
+	title = {BigFoil.com Airfoil Database},
 	year = {2022},
+	url = {https://www.bigfoil.com}
 	note = {\url{https://www.bigfoil.com}},
 }
 %%%% use our representation
