### 2023-09-19 d3f2ca6
minor edit to paper.bib
