+++ paper/paper.bib
@@ -185,7 +185,7 @@
 	author = {Quayle, Michael},
 	title = {BigFoil.com Airfoil Database},
 	year = {2022},
-	url = {https://www.bigfoil.com}
+	url = {https://www.bigfoil.com},
 	note = {\url{https://www.bigfoil.com}},
 }
 %%%% use our representation
