### 2023-09-19 NREL/G2Aero#6 @olgadoronina
Made changes in the .bib file to fix the issue.
