### 2023-09-19 #5408 @olgadoronina
@vissarion, I made some minor edits to fix the [NREL/G2Aero#6](https://github.com/NREL/G2Aero/issues/6)
