+++ paper/paper.bib
@@ -137,7 +137,7 @@
 }
 
 @article{geuzaine2009gmsh,
-	title={Gmsh: A 3-D finite element mesh generator with built-in pre-and post-processing facilities},
+	title={Gmsh: A 3-{D} finite element mesh generator with built-in pre-and post-processing facilities},
 	author={Geuzaine, Christophe and Remacle, Jean-Fran{\c{c}}ois},
 	journal={International journal for numerical methods in engineering},
 	volume={79},
@@ -173,7 +173,7 @@
 
 @misc{wisdem,
 	title = {WISDEM{\textregistered} (Wind-Plant Integrated System Design and Engineering Model), Version v3.4.1},
-	author = {Dykes, Katherine and Ning, S. Andrew and Scott, George and Graf, Peter and USDOE Office of Energy Efficiency and Renewable Energy},
+	author = {Dykes, Katherine and Ning, S. Andrew and Scott, George and Graf, Peter and {USDOE Office of Energy Efficiency and Renewable Energy}},
 	doi = {10.11578/dc.20211208.2},
 	url = {https://www.osti.gov/biblio/1834705}, 
 	year = {2021},
