### 2023-09-20 #5408 @danielskatz
👋 @olgadoronina - I'm the track editor, and having proofread the paper, found a couple of small issues, as in https://github.com/NREL/G2Aero/pull/7 - please merge this, or let me know what you disagree with, then we can proceed to acceptance and publication.
