### 2023-09-20 #5408 @danielskatz
Congratulations to @olgadoronina (Olga Doronina) and co-authors on your publication!!

JOSS is dependent on volunteers and couldn't be successful without your efforts

(the paper does not yet load for me, so I will keep this issue open until it does)
